@@ -2,19 +2,12 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="..\..\..\..\..\..\Data\Migrations\20220717130702_PlanBetter.cs" Link="20220717130702_PlanBetter.cs" />
-    <Compile Include="..\..\..\..\..\..\Data\Migrations\20220717130702_PlanBetter.Designer.cs" Link="20220717130702_PlanBetter.Designer.cs" />
-    <Compile Include="..\..\..\..\..\..\Data\Migrations\20220717130736_AddMyPlanBetter.cs" Link="20220717130736_AddMyPlanBetter.cs" />
-    <Compile Include="..\..\..\..\..\..\Data\Migrations\20220717130736_AddMyPlanBetter.Designer.cs" Link="20220717130736_AddMyPlanBetter.Designer.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="5.0.17" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="5.0.17">
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="6.0.7" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="6.0.7">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
